made the spline generate a mesh
