--- SpaceGame/Assets/Mail.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5caa7a762c963f04ab1c38e0a3b004d2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
