--- SpaceGame/Assets/MailText.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 455f3e88308b94f4b8184fa87a4525f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
