+++ SpaceGame/Assets/Prefabs/BorderCollider.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 978e713d6a7c3144bb371c6834ed064b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
