+++ SpaceGame/Assets/Scripts/Editor/SplinePrefabPlacerEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: db844e80903143ee8c9948e8df4d877e
+timeCreated: 1592899712
