+++ SpaceGame/Assets/Scripts/MailCounter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c2d91fbd32ef2d40aa46822786d0ecc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
