+++ SpaceGame/Assets/Scripts/SplinePrefabPlacer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 59252191cbe8ebe4ca6a81746c5ac822
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
